Removed Client dependency on Core
