+++ src/Client/Models/JobSchema.cs
@@ -1,4 +1,4 @@
-namespace Runly.Models
+namespace Runly.Client.Models
 {
 	public class JobSchema
 	{
