+++ src/Client/Models/Package.cs
@@ -3,7 +3,7 @@ using Newtonsoft.Json.Converters;
 using System;
 using System.Collections.Generic;
 
-namespace Runly.Models
+namespace Runly.Client.Models
 {
 	public class Package
 	{
