+++ src/Client/Models/Pagination.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace Runly.Models
+namespace Runly.Client.Models
 {
 	public class Pagination<T>
 	{
