+++ src/Client/QueryString.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace Runly
+namespace Runly.Client
 {
 	public static class QueryString
 	{
