+++ src/Client/RunClient.cs
@@ -1,5 +1,5 @@
 using Newtonsoft.Json.Linq;
-using Runly.Models;
+using Runly.Client.Models;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -8,83 +8,19 @@ using System.Net.Http.Headers;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Runly
+namespace Runly.Client
 {
 	public interface IRunClient
 	{
 		Task<Pagination<Run>> Search(string organization, RunSearch searchParams = null);
 		Task<IEnumerable<Run>> GetActiveRuns(string organization, string environment);
-
 		Task<Run> GetRun(string organization, Guid runId);
-		Task<TConfig> GetConfig<TConfig>(string organization, Guid runId) where TConfig : Config;
-
+		Task<string> GetConfig(string organization, Guid runId);
 		Task<IDictionary<RunLogType, RunLogInfo>> GetLogInfo(string organization, Guid runId);
 		Task<Stream> DownloadLog(string organization, Guid runId, RunLogType type);
-
-		Task<Run> Enqueue(string organization, string environment, Config config);
-		Task<Run> EnqueueFromTemplate(string organization, string environment, string template, object configToMerge = null, bool scheduled = false);
-
+		Task<Run> Enqueue(string organization, string environment, string config);
 		Task<Run> Requeue(string organization, Guid runId);
 		Task<Run> Cancel(string organization, Guid runId);
-	}
-
-	public static class RunsClientExtensions
-	{
-		public static Task<Run> Enqueue(this IRunClient client, string organization, string environment, string jobType, string version = null, string package = null)
-		{
-			return client.Enqueue(organization, environment, new Config
-			{
-				Job = new JobConfig
-				{
-					Type = jobType,
-					Version = version,
-					Package = package
-				}
-			});
-		}
-
-		public static Task<Run> Enqueue<TJob>(this IRunClient client, string organization, string environment, string version = null, string package = null)
-			where TJob : class
-		{
-			return client.Enqueue(organization, environment, new Config
-			{
-				Job = new JobConfig
-				{
-					Type = typeof(TJob).FullName,
-					Version = version,
-					Package = package
-				}
-			});
-		}
-
-		public static Task<Run> Enqueue<TConfig>(this IRunClient client, string organization, string environment, string jobType, TConfig config = null, string version = null, string package = null)
-			where TConfig : Config
-		{
-			Config cfg = config ?? new Config();
-			cfg.Job = new JobConfig
-			{
-				Type = jobType,
-				Version = version,
-				Package = package
-			};
-
-			return client.Enqueue(organization, environment, cfg);
-		}
-
-		public static Task<Run> Enqueue<TJob, TConfig>(this IRunClient client, string organization, string environment, TConfig config = null, string version = null, string package = null)
-			where TJob : class
-			where TConfig : Config
-		{
-			Config cfg = config ?? new Config();
-			cfg.Job = new JobConfig
-			{
-				Type = typeof(TJob).FullName,
-				Version = version,
-				Package = package
-			};
-
-			return client.Enqueue(organization, environment, cfg);
-		}
 	}
 
 	public class HttpRunClient : IRunClient
@@ -133,8 +69,7 @@ namespace Runly
 			return await response.Content.ReadAsAsync<Run>();
 		}
 
-		public async Task<TConfig> GetConfig<TConfig>(string organization, Guid runId)
-			where TConfig : Config
+		public async Task<string> GetConfig(string organization, Guid runId)
 		{
 			var req = new HttpRequestMessage(HttpMethod.Get, $"/{organization}/runs/{runId}/config");
 			req.Headers.Authorization = await tokenProvider.AcquireAuthHeader();
@@ -142,7 +77,7 @@ namespace Runly
 			var response = await api.SendAsync(req);
 
 			await response.EnsureSuccess();
-			return await response.Content.ReadAsAsync<TConfig>();
+			return await response.Content.ReadAsAsync<string>();
 		}
 
 		public async Task<IDictionary<RunLogType, RunLogInfo>> GetLogInfo(string organization, Guid runId)
@@ -175,26 +110,10 @@ namespace Runly
 			return await response.Content.ReadAsStreamAsync();
 		}
 
-		public async Task<Run> Enqueue(string organization, string environment, Config config)
+		public async Task<Run> Enqueue(string organization, string environment, string config)
 		{
-			var req = new HttpRequestMessage(HttpMethod.Post, $"/{organization}/environments/{environment}/runs/").WithJsonContent(config);
-			req.Headers.Authorization = await tokenProvider.AcquireAuthHeader();
-
-			var response = await api.SendAsync(req);
-
-			await response.EnsureSuccess();
-			return await response.Content.ReadAsAsync<Run>();
-		}
-
-		public async Task<Run> EnqueueFromTemplate(string organization, string environment, string template, object configToMerge, bool scheduled)
-		{
-			var jobj = JObject.FromObject(configToMerge ?? new { }, ConfigWriter.Serializer);
-			jobj["template"] = template;
-			jobj["scheduled"] = scheduled;
-			string json = jobj.ToString();
-
 			var req = new HttpRequestMessage(HttpMethod.Post, $"/{organization}/environments/{environment}/runs/");
-			req.Content = new StringContent(json, Encoding.UTF8, "application/json");
+			req.Content = new StringContent(config, Encoding.UTF8, "application/json");
 			req.Headers.Authorization = await tokenProvider.AcquireAuthHeader();
 
 			var response = await api.SendAsync(req);
