+++ src/Runly/Internal/RunAction.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.Extensions.Logging;
-using Runly.Models;
+using Runly.Client.Models;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
