+++ src/Runly/ResultsChannel.cs
@@ -2,7 +2,8 @@ using Microsoft.AspNetCore.Http.Connections.Client;
 using Microsoft.AspNetCore.SignalR.Client;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using Runly.Models;
+using Runly.Client;
+using Runly.Client.Models;
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
